requirements.txt
feedparser>=6.0
requests>=2.31
Jinja2>=3.1
PyYAML>=6.0
weasyprint>=60.0
pytest>=8.0
